requirements/requirements-base.txt
six>=1.10
djangorestframework>=3.15.0
django-rest-swagger>=2.2.0
django-filter>=1.1.0
